1/n: buck bxl: add flag to control building generated buck targets
Summary:
The `elp.bxl` prelude file has just been updated to take an extra boolean flag to trigger invoking a build of the targets that are generated.

This diff provides a matching flag in the ELP CLI arguments, called `--buck-generated`.

If set, this flag implies the usage of buck BXL, and is plumbed though into the project discovery to add the `--build_generated_code true` to the invocation of `elp.bxl`. It also updates the progress report to alert the user that generation is being done.

Reviewed By: TheGeorge

Differential Revision: D71904533

fbshipit-source-id: a3fb7918f895a5b25fe57848b99821c91933c2c5