2/n: buck bxl: Make a buck2 test project that generates a file
Summary:
Add an ELP-local test to show
- that the generation step is invoked when the flag is set
- that the `srcs` field in the generated target is now populated.

Note: We are working on the basis that generation will only populate source files for targets that already exist, not create new ones.

This opens the door to a two-step process, first loading the model without generating, then generating the output and loading the new files only, without invalidating the rest.

Reviewed By: TheGeorge

Differential Revision: D71905205

fbshipit-source-id: 7af1f120836ec7be5434f5ca08ace2c3e440aa4a