Add logging for changes to watched files in IDE mode
Summary:
We are seeing that sometimes we have a stale state in the IDE, particularly aafter branch switches.  A likely cause is that we are not properly detecting changes to files on the file system.

To assist in debugging this situation, this diff adds a local log file which simply records the files that ELP gets told have changed, either from the client (VS Code) or the VFS file watcher.

If the IDE ends up in an inconsistent state, we can check the known changed files against this.

Technicalities.

It installs an additional file logger, configured at info level with a specific target name, and then logs against this combination. This means that if the overall log level is dropped to info these logs will show up there too.

It also uses the `humantime` crate, already in our crate graph from `env_logger` to give a time stamp to the file logger output to better understand which changes are recent.

Reviewed By: jcpetruzza

Differential Revision: D71977224

fbshipit-source-id: 9a5a81ef7bd03bd0f3123e841edcb8d4ac2847aa