Preserve spacing during conversion to Markdown
Summary:
Even if leading spaces are ignored in Markdown, eDoc-style comments may contain whitespaces that the user would like to preserve during the Markdown conversion. Think to a ASCII diagram.

Here we change the conversion logic to preserve leading spaces, trimming the common sub-set of whitespaces.

We also adjust the formatting of unknown tags to use Markdown h3 titles, rather than bold, for simplicity.

Finally, we simplify the logic of the "return" tag conversion by reusing the common one.

Reviewed By: alanz

Differential Revision: D72172379

fbshipit-source-id: 5f875d50afded631b582d36cd2091c0d2903d1d4