Handle @see tag during conversion to Markdown
Summary:
The `see` tag can be used in Edoc-style comment to point to a reference.
So far, we've been treating the tag as a generic one, but here we specialize the conversion, by prepending the "See" text and wrapping the eventual reference in backquotes.

Validating the reference is out of scope for this change (the reference syntax is basically equivalent in old-style and new-style edoc for our purposes)

Reviewed By: michalmuskala

Differential Revision: D72226821

fbshipit-source-id: a2675e4e8b013a5c8a2a69eb9c38fb840915ce2a