BE: make testing fixes more robust
Summary:
Our tests can check for a fix, based on a diagnostic.
We indicate in the test fixture the location that the diagnostic and fix should occur.

This diff updates the `check_fix` function to filter the diagnostic fixes to limit them to ones spanning the required location. This makes them more robust if the diagnostics appear in a differnent order, such as edoc ones.

Reviewed By: TD5

Differential Revision: D72310416

fbshipit-source-id: 4b23fec51d9ab49697bb7ada9e5f006b4ed9b9b8