ssr: Fix multiple matches from different forms
Summary:
Once we have made all the SSR matches in a file, we run a process to de-duplicate them.  This was using an id for the match that was only valid within the context of a specific body, and so the same index value could occur for a match from a different body.

This diff fixes the problem by storing the matches by `BodyOrigin` as well, so we do have a globally unique index for each match.

Reviewed By: robertoaloi

Differential Revision: D72316768

fbshipit-source-id: 33e50fdc9a85a6ee2d6c2cb3466c1816f24484d2