Handle the case where a function is both marked as hidden and contains docs
Summary:
# Context
We have occurrences in the codebase where a function contains documentation and it's also marked as "private".
# Problem
While EEP 59 supports having both a string documentation and the hidden attribute, [a bug in OTP](https://github.com/erlang/otp/issues/9672) causes the code to fail compilation.
# Solution
Until the bug is resolved, comment out the hidden tag if the function contains actual documentation, linking to the bug report.

Reviewed By: jcpetruzza

Differential Revision: D72297138

fbshipit-source-id: 1b8e8ffc5691f5bd92167b13c15f791b66caf12b