Fix support for @see in combination with @end
Summary:
Sometimes the `end` tag (whose purpose is to end the previous tag) is placed directly after a `see` one, and intended to (incorrectly) close a previous `doc` block.
In this case, during the conversion to Markdown, we could take an extra newline as part of the reference.
Here we take into account this cornercase, by improving the reference extraction.

Reviewed By: jcpetruzza

Differential Revision: D72261572

fbshipit-source-id: 86fa0b8e5b8328b122035467968b30d122e32c59