IDE: Always show eqwalizer diagnostics
Summary:
# Problem

In the server, we ask eqwalizer for both diagnostics and types, and only return a value if we have a result.

But we have two data sources, and used `?` on both, so if one did not return a result we do no return a value at all.

The types query only returns a value if it processes at least one function. So for a file having an erroneous type definition, but no function, it will not report the eqwalizer diagnostic

# This diff

We fix it by separately requesting both data sources, falling back to a default empty list. If either has something to report, we do so.

Reviewed By: TD5

Differential Revision: D72379193

fbshipit-source-id: 37c9c2c4aa0771bebea7e11603f71c5197edb1e9