eqwalizer: intern all strings in eqwalizer AST, take2
Summary:
Attempt2. This time with more fixes to avoid instability in transitive validation check - see diffs earlier in the stack

With this change we intern all strings in eqwalizer AST. This is done by switching from `SmolStr` to `Ustr`. I'm not sure how much of an impact interning everything has (given that SmolStr doesn't allocate for strings shorter than ~20 characters). However, we include those strings in many data structures and `SmolStr` is 24 bytes in size, while `ustr` string is only 8 bytes - this shrinks pretty much all data structures that eqwalizer uses.
In particular, things like `Id` are shrunk from 32 bytes to 16, and `RemoteId` from 56 to 24.

With some testing with `/usr/bin/time -v` this shrinks the max ELP size by ~1GB from ~8.9 to ~7.9, or ~11%.

The trade-off is that once allocated, those strings are never freed. However, the identifiers in AST are fairly stable, and I don't think this should be an issue in practice - we don't reload a lot of code for eqwalizer - we convert only once the code passes `erlc` linter, so it's mostly valid.

Why not salsa interning? It gives us back an ID we have to use back with salsa to get the value - this is API-incompatible with how serde works, so we wouldn't be able to "just" use it here.

Reviewed By: alanz

Differential Revision: D71540337

fbshipit-source-id: 4bde3e5dd58d17c9c241b150f2d13b87c9902d5d