memory stats: include number of loaded files in the telemetry
Summary:
ELP is now reporting open files for the telemetry, but we also load the entire project into salsa/vfs on startup, and do an initialy analysis on each file.

So report the number of loaded files too.

Note: we use the current highest seen `FileId` index value for this count, as every loaded file has one, and it increments monotonically.

Reviewed By: jcpetruzza

Differential Revision: D72391467

fbshipit-source-id: 1fa151f6c6c338550d9b4510c3e05e63851a1ca2