1/n: Update HIR for nominal types
Summary:
Match the changes made in the prior diff to introduce nominal types in the tree-sitter-erlang grammar.

Note that because we already support normal and opaque type attributes, this is a straightfoward addition of a new variant to an enum.

The tests are updated in the next diff, to make it clearer.

This also renames `map_expr` to `map_expr_for_completion` to clarify why an opaque receives a `None` result

Reviewed By: ilya-klyuchnikov

Differential Revision: D72445311

fbshipit-source-id: e27f3535384b3fd35a1a1b191eb6b2e5ec06d063