Do not report undocumented_function diagnostics for simple functions
Summary:
Functions with a trivial implementation have less need to be documented. This diff brings in the idea of excluding "simple" functions from the "undocumented function" diagnostic. The idea is to focus developers documentation effort to the functions that would benefit the most from it.

The heuristic used here is trivial: if a function body (including comments) does not exceed 5 lines (across multiple clauses) consider the function "simple" enough, not requiring documentation. Of course a developer could still decide to document it.

This logic is just a starting point and can be later extended to something more meaningful as we go.

Ideally, we'd made the complexity parameters (e.g. the min number of lines) configurable, but right now we don't really have a way to provide linter configuration externally. That is out of scope for this change.

Reviewed By: alanz

Differential Revision: D72453919

fbshipit-source-id: 468b9d46b19b38f5d08a174a3fb8c08ae1142936