Bump rust-analyzer dependency to 2024-08-12
Summary:
The rust-analyzer dependency is currently 2024-07-29.  It turns out that a few days after that an important performance improvement landed, which does parallel loading for file in VFS. This has a significant impact on initial startup time.

This  diff does the bare minimum to move to the first subsequent release that had this feature, from [2024-08-12](https://github.com/rust-lang/rust-analyzer/releases/tag/2024-08-12).

The key change is [this one](https://github.com/rust-analyzer/rust-analyzer/pull/17771) "load VFS config changes in parallel."

Reviewed By: jcpetruzza, michalmuskala

Differential Revision: D72724277

fbshipit-source-id: eac8f371acfa548df0c51fcf2c6c49fe5417f36f