buck bxl: tag buck prelude modules as third party
Summary:
They are not part of our code base, so should not have diagnostics reported on them.

The `elp.bxl` process takes the included targets, and marks them as `apps`, then blows out their dependencies, and marks that set as `deps`.  These include *all* dependencies for the targets.

Inside ELP, we only report on "current project" files.  These are ones explicitly marked as apps, so you can add a third party or prelude target if you need to.

We then flag any of the deps targets that are in prelude or third party as third party. Which is what this diff does.

Reviewed By: TheGeorge

Differential Revision: D72790040

fbshipit-source-id: 323b72e701fb6aab1e6db3e008e7efcbc9fc3991