Report detailed project loading telemetry in the IDE
Summary:
One of the key user-experience items is how long it takes the language server to be available on startup.

To track this, we have introduces detailed user feedback by means of reporting a spinner message as each step completes, visible in the UI.

This diff takes it one step further, and records detailed timing for each step, which is then reported in a telemetry message once the loading is complete.

This will allow us to track startup as a whole, as well as focus on the individual steps for optimisation.

Reviewed By: jcpetruzza

Differential Revision: D73030961

fbshipit-source-id: b1db9db677379c3a890ffedad109c2815a5b1a48