IDE: tweak spinner feedback message
Summary:
In the IDE, we now provide UI feedback of the detailed project loading steps.

One of the first ones is to load the project manifest, i.e. the relevant buck or rebar config.

This diff changes the wording for this step, so indicate it is loading the config, not the project as a whole, because there are still a number of steps after it before the project is loaded.

Reviewed By: jcpetruzza

Differential Revision: D73027165

fbshipit-source-id: fd966d4711f3a08a1aaf9b635ff67cf34d66b875