Decouple IDE from CLI severity
Summary:
In some cases (e.g. in the `debugging_function` linter) we may want different severities for IDE and CLI (e.g. in a CI context).
Here we introduce an additional, optional, `cli_severity` field which allows to specify an additional severity for a diagnostic. The additional severity is only used if the `--use-cli-severity` argument is provided by the CLI.

Reviewed By: alanz

Differential Revision: D72624965

fbshipit-source-id: becec0977f057eb89188e25cd5c7a72b7dfd3548