Typed expressions
Summary:
Implementation of typed expressions on ELP's side.

The essence of this diff is the IPC protocol which deserializes the message `ValidateType(ExtType)`, expands and transitively checks the type, then answers either `ValidatedType(Type)` or `InvalidType(Invalid)` depending on the result of the checks.

The macro setup and testing will come in the next diff.

Reviewed By: ilya-klyuchnikov

Differential Revision: D71807120

fbshipit-source-id: 38a9c923a19b1e5a2823c1801dc220512701b96d