cargo fix --edition
Summary:
This is step two in the [rust edition transition guide](https://doc.rust-lang.org/edition-guide/editions/transitioning-an-existing-project-to-a-new-edition.html)

I took a two-step process, first running

```
cargo fix --edition
```

(Output in P1787020079)

This resulted in many `if let Some(..) .. {..} else {..}` being converted to match statements.

In the second step, I undid these, as the scope changes do not have an impact.

The matching of expressions in macro definitions [has changed](https://doc.rust-lang.org/edition-guide/rust-2024/macro-fragment-specifiers.html), so the migration has converted types of `expr` to `expr_2021`.

I left these as is, so we can update them separately, if needed.

Reviewed By: robertoaloi

Differential Revision: D73108732

fbshipit-source-id: 99bd80792cdea146938e2808ce13233e67c26724