BE: Add plumbing for workspace-level lint configuration
Summary:
Rust 1.74 introduced [lint configuration through cargo](https://blog.rust-lang.org/2023/11/16/Rust-1.74.0/#lint-configuration-through-cargo)

With a view to ending up clippy-clean, we add a [workspace lint configuration](https://doc.rust-lang.org/stable/cargo/reference/workspaces.html#the-lints-table) at the workspace root and enable it in all crates.

For the initial test we allow the use of things that could be replaced by the `matches!` macro, which is clippy terminology for turning that lint off.

Reviewed By: robertoaloi

Differential Revision: D73120037

fbshipit-source-id: 081a30f7984d2cad4b3e0702f7c313694ebc1be3