Fix go to def for multi var when on a definition
Summary:
The logic for processing `to_def` on a variable when the given variable was already a definition was incorrect, in that it assumed there can only be one definition site for a variable, and had an assert for this case.

In Erlang a given variable can have multiple definition sites, from resolving matches, or occurring in legs of a case expression.

This diff preserves the old behaviour if there is indeed a single definition site and we are on it, but provides multiple definitions otherwise.

Reviewed By: VLanvin

Differential Revision: D73507506

fbshipit-source-id: 811b9d5fdc1531898387a0baab3454d836d0ef9e