Add `--no-buck-bxl` command line argument, and make bxl default
Summary:
This diff flips the default buck project model to be using the `elp.bxl` query.

It also provides a new command line argument `--no-buck-bxl` to use the old model, if required during the transition.  This is used in some of the existing tests to keep them passing with the current expected output, so demonstrates that it works properly.

The default is still not to ask `elp.bxl` to build generated code, this is is controlled by `--buck-generated`

Reviewed By: robertoaloi

Differential Revision: D73499601

fbshipit-source-id: 8b7c39db06f960f2b764bb7da7aa1d4c34cd3a86