BE: Improve naming around getting the range of a matched function
Summary:
`find_call_in_function` invokes a callback with a `MatchContext` as an argument.

This includes two range fields, one of the whole match including arguments, and one of the `MF` part only, but only if it is not part of a macro expansion.

There is also a utility function to always provide a range, using the targeted `MF` if it is at the surface level, or the entire macro otherwise.

Change the naming and comments around these to make the situation clearer.

Reviewed By: robertoaloi

Differential Revision: D73585276

fbshipit-source-id: 2526e91106d3f50be7705fff8d0bc66999eff50b