fold: apply MacroStrategy to Pat::MacroCall as well
Summary:
Our fold traversal of an `Expr::MacroCall` and `TypeExpr::MacroCall` honour the `FoldBody.macros` flag, traversing the macro arguments instead of the expansion if `MacroStrategy::DoNotExpand` is provided.

But this was not implemented for `Pat::MacroCall`.

This diff implements it for `Pat::MacroCall`, and also updates the tree_print logic when initialised with a strategy to make it clear if an expansion will be seen or not.  The intent is that a tree_print with strategy provides the same view that a fold with that strategy would have.

Note: we simply process the expansion for `Term::MacroCall` at present, as the strategy-based approach has not been needed up to now, Terms appear in very few contexts.

Reviewed By: robertoaloi, TD5

Differential Revision: D73657708

fbshipit-source-id: 65be482fa38cd18fe0e70158e1e6db3bb02d6ca1