Add --app option to lint command
Summary:
It is currently possible to invoke the `lint` command on a full project or a single module, but there's no way to do so on a single app. This change solves this problem by adding a `--app` option.

This is in line with what `eqwalizer` does via the `eqwalize-app` command.

If both a `module` and an `app` are specified but the module does not belong to the app, an error is raised.
When only an `app` is specified, we currently do not validate the app name. This is because there's currently no API to efficiently check whether an app exists. We could verify the number of processed files instead, and return an error or a message if no modules are processed. An even better approach would be to refactor the code to have a `process_many` instead, at which point we could have a centralized check. But this is not essential for now, so leaving it as a potential follow-up.

Reviewed By: alanz

Differential Revision: D73658452

fbshipit-source-id: 9b4f81585465eea264ecdc4be4a159105bf33147