Update include_lib generation for vagaries of buck2 generated files
Summary:
When using a buck2 project, elp.bxl is able to invoke the generate step to generate any artifacts required.  It does *not* invoke the build step.

The build step in addition to the generation arranges for a mapping of included files in the project into a directory structure in the buck out directory corresponding to a normal (rebar) Erlang application. It does so by means of symlinks.

We want to be able to work with these files in ELP in the generated locations, which are also in buck out but not with the mapping symlinks applied.

This diff allows for that, by adding a fallback case to the `include_lib` processing that strips out the extra "include" part of the path for an application, and looks for the requested file directly where it is generated.

Reviewed By: robertoaloi

Differential Revision: D73785465

fbshipit-source-id: 19476481305a3eda6994ed3c621ae2a11d069035