simpler contractivity (2)
Summary:
Now, - after the prev diffs, the only reason that we have to jungle with `Result<..>` is this piece of code

https://www.internalfb.com/code/fbsource/[44c4c604fa08f0623783579cb82b1dbd943d503c]/fbcode/whatsapp/elp/crates/eqwalizer/src/ast/contractivity.rs?lines=340-343

While it may happen according to the types, it cannot happen in practice. Why?

Consider that a remote type referenced this way:
```erlang
-module(m2).

-type t2() :: m1:t1().
```

What can go wrong?
- Module m1 doesn't exist
  - In this case expansion phase for `m2:t2()` would fail and we won't be checking contractivity of `m2:t2()` - since we check it only for types that successfully passed the previous phases
- Module m1 does exist, but is not parseable. Again the expansion phase validates that all the referenced types (from other modules) exist and are exported. So, the expansion phase for `m2:t2()` would fail.

 ---

Migrating from defensive code to "let it crash if we are here" allows to remove a lot of Error-handling boilerplate.

Reviewed By: michalmuskala

Differential Revision: D73781233

fbshipit-source-id: fc0382240999f3f0249ecaf63a165dcfef35df70