Handle @deprecated tag in conversion to Markdown
Summary:
We currently handle the `deprecated` tag in a generic way during the conversion to Markdown. This means the tag is converted to a `h3` section.
Since the Erlang [metadata](https://www.erlang.org/doc/apps/edoc/chapter.html#function-tags) define a deprecated keyword, use that one instead.

Since the keyword can also be used as part of `moduledoc`, generalize the prefix used.

Reviewed By: jcpetruzza

Differential Revision: D74056601

fbshipit-source-id: 078ddf52af81706b5791a0638594b9ed56638341