4/n: Use Buck UI Url in server when the bxl query fails
Summary:
The final step is to present the failure in the IDE UI too.

Up to now we sent an error notification, with a long message in it, which was incomprehensible because it is not presented as formatted test.

Instead, we present the simpler representation used in the CLI output, but if we have a URL to open the buck session, we present it via a button. Clicking it opens the web page.

Reviewed By: Balajiganapathi

Differential Revision: D73996021

fbshipit-source-id: 4f965708e8eb6a6e0fbd59584271e38095ee4e9b