2/n: Introduce BuckBxlQueryError
Summary:
Intoduce a new type specifically to carry information about the failing buck bxl query.

This is intially populated with the three items we had before (command, failure reason and details).

The key point is we now have a `Display::fmt` implementation for it, so can tweak what we gather and show in one place. To start with this replicates the existing error message.

We simply insert the type, and check that no existing tests break because of it

Reviewed By: Balajiganapathi

Differential Revision: D73995460

fbshipit-source-id: f137bdefd2aeba94a158172bc56f78e1164c7e31