BE: BodySourceMap has a FileId in the result, validate it on use
Summary:
In HIR, we have `Body` and `BodySourceMap` that track the HIR position-independent ast and a mapping from each item to the original source.

This mapping returns both a `FileId` and `TextRange` for each item, which can be different from the originating file if the contents are the result of macro expansion.

Up to now we have been ignoring the `FileId`.

This diff changes the API to always return a `FileRange`, which contains both the `FileId` and `TextRange`, and explicitly checks for a match before generating a diagnostic, or other usage.

Reviewed By: robertoaloi

Differential Revision: D74242563

fbshipit-source-id: 35c5b96c057118414f3c4949a0f32a11ee4763cb