BE: clarify inverse relationship for generated include_lib
Summary:
There are some subtleties around the include_lib processing for generated or mapped buck targets.

We currently deal with them in the fallback processing for include_lib, and also when generating the `unspecific_include` (W0037) quick fix.

To simplify future maintenance, this diff puts the two inverse functions next to each other, so we stand a better chance of updating both if anything changes.

Reviewed By: TheGeorge

Differential Revision: D74400170

fbshipit-source-id: fd5d3a3bc4b2b8233fd228fee83e8fba4e4948c7