Lower moduledoc and doc metadata attributes
Summary:
The goal of this stack is to be able to extract parameter information (and eventually other information) from the [EEP59 function doc metadata](https://www.erlang.org/doc/system/documentation.html#doc-metadata), so that they can be used in features such as *signature help*.

In this first step, we specialize the *lowering* of `-moduledoc` and `-doc` attributes. If the attribute contains a map, we lower it as, respectively, `ModuleDocMetadataAttribute` or `DocMetadataAttribute`. If not, we lower it as the usual `ModuleDoc` and `DocAttribute`.

We keep track of the latest function doc metadata while constructing the `def_map`. This allows us, similarly to what happens with the regular doc attribute, to associate the doc metadata attribute id with the function definition, for easy retrieval.

We also patch the *folding range* feature to take into account the new specialization.

We finally tweak the `add_doc` quick-fix to take the doc metadata into account. In the process, we move some computation from the quick-fix to the `FunctionDef`, to make it easier to discover and reuse.

Reviewed By: alanz

Differential Revision: D74313724

fbshipit-source-id: 19b3eac6e7da07d8f9cf32c542996e7e512981e8