Improve maps:find diagnostic to handle more cases
Summary:
Improves the `maps:find` rewriting diagnostic to handle some common variants seen in the wild, such as including a guard and having two potential values matched.

We stop expanding macros to avoid a rewrite losing the expanded macro, which may be intended as a form of documentation. Similarly, we don't apply rewrites which might clobber comments.

**N.B.** We just tackle common examples here, rather than trying to approach full generality. As SSR grows in power, we can both reduce boilerplate and handle more cases.

Reviewed By: alanz

Differential Revision: D74326852

fbshipit-source-id: 861365a803200895f76e3333bb4adebed9ba87eb