Unmark heavier diagnostics as simplification rules
Summary: We remove the "simplification rule" tag from some of the heavier diagnostics to keep fixes targetted. For example, if you use `elp lint` to power a code-modding system, and you want to migrate a certain kind of function call, you probably don't want to drag in a rewrite of a `lists:fold` too.

Reviewed By: jcpetruzza

Differential Revision: D74477202

fbshipit-source-id: 499a6b70c6160ca07ad1de8d2ef8ca0c54492c97