Apply gatekeeper values in ELP, not the extension
Summary:
Internally, we have a configuration system where we can set flags called gatekeepers.  These values get inserted into the `initialize` message by the internal extension.

Up to now we have been harmonising the config values in the extension, so the GKs take precedence over the config value, but this is leading to performance issues and race conditions if the user has the config file open at the same time.

Instead, we keep a copy of the startup gatekeeper values sent in the `initialize` message, and apply them to the config we receive from the client whenever that is updated.

Reviewed By: robertoaloi

Differential Revision: D74466115

fbshipit-source-id: a61e3c9c18f313dfead8ba2d293f1b6e18d0b023