Add --no-buck-generated argument, and make generated default
Summary:
As we did for the `--buck-bxl` flag, we add a `--no-buck-generated` flag and flip the default to always generate code when invoking buck.

This means existing usage of `--buck-generated` becomes a no-op, and we can clean them up in subsequent diffs once this change has propagated to all CI infra.

Reviewed By: TD5

Differential Revision: D74574966

fbshipit-source-id: dd08969ebe5a597c5ce002dc2c384227c060ba7f