Add undocumented_module diagnostic
Summary:
# Context

Currently we track the list of undocumented functions via the ELP diagnostic code `W0040`.

# Problem

Many functions belong to internal modules that do not require documentation. Due to the amount of undocumented functions, developers cannot focus on the important ones.

Furthermore, sometimes a module overview is even more important of specific function documentation.

# Solution

Track undocumented modules in a similar way to undocumented functions. This will allow developers to quickly identify modules that do not require documentation. By adding a `-moduledoc false`, functions belonging to the same module will automatically be marked as not requiring documentation. For the time being, the linter is disabled by default.

Reviewed By: alanz

Differential Revision: D74582638

fbshipit-source-id: 272d30b0ebf70d1f8761ea467ff81fd4faa8caa7