Add garbage collect linter
Summary:
While garbage collection happens automatically, it can manually be triggered via
the `erlang:garbage_collect/0,1,2`. This function forces an immediate garbage
collection of the executing process. The function is not to be used unless it
has been noticed (or there are good reasons to suspect) that the spontaneous
garbage collection will occur too late or not at all. Improper use of this
function can seriously degrade the performance of the system.

Add a linter that warns the user about usages of explicit garbage collection.

Reviewed By: alanz

Differential Revision: D73923933

fbshipit-source-id: f708d4e2dc8141167f05506ef43c8ba8a717bc48