Restrict range of undefined type error to name only
Summary:
For the "undefined type" diagnostic (L1295), Erlang/OTP returns the range of the whole `export_type` attribute. Refine it to only cover the problematic type.

For simplicity, apply a small tweak to the linter so that, instead of returning a number of underscores - one per parameter -, just return the arity.

Note: this is modeled after the "undefined function" one, but both can probably be migrated to operate at hir level, instead.

Reviewed By: jcpetruzza

Differential Revision: D75052714

fbshipit-source-id: 4ef03ea015a57d11bddfe6a04b905b1035bbbb28