Do not warn about unused ct.hrl header
Summary:
It is common to include the `ct.hrl` header in test suites without using it. This can happen, for example, when using an editor template for a new test suite.

Here we add the ability to silent the "unused header" diagnostic for some specific headers, `ct.hrl` being one of them.

We could limit this to only test suites, but that would make computing the diagnostic more expensive (we need to extract the module name, check for the suffix, etc). I suspect it's not worth the trouble, so just ignoring it for all files.

Reviewed By: TD5

Differential Revision: D75065398

fbshipit-source-id: ed40045c718a3bc1d440f1d2d0ccc8db013e18db