Make optimising diagnostic fixes more defensive
Summary:
In particular, where applicable, be very careful not to risk affecting the positioning of comments or to end up applying changes to included header files. This makes it safer to automatically apply these fixes where they apply.

I've not factored the new sensibility checks out any further, since different diagnostic fixes have different properties, so they are not all necessarily the same and this is an on-going refinement.

Reviewed By: alanz

Differential Revision: D75073883

fbshipit-source-id: 7eebb0f33eafb6e47e073d48f1c01de03d48af81