Add release-thin profile
Summary:
In D74474845, we introduced some heavy optimizations for the `release` profile of ELP. This led to improved performance, but slower compilation times.

This diff introduces an extra profile (`release-thin`), that developers can use as a compromise between decent build times and decent performance.

To enable it:

```
cargo build --profile release-thin
```

Reviewed By: TD5, michalmuskala

Differential Revision: D75072411

fbshipit-source-id: ea8c11a6fb040dc1bc123ef7a31eb060c806caf6