more precise location for unused type warning
Summary:
The quick fix (see export_type.rs) is available for the range of the type alias name, - while the diagnostics for unused type from erlang_service was shown for the whole type definition - so it seemed that the quick fix is not available for the warning.

Trimming the range of the warning down to the name range.

Reviewed By: alanz

Differential Revision: D75066457

fbshipit-source-id: 47fcc02e678fe7a916b79eb5eb3a65b72c16fcc0