9/n: Keep local include dirs in AppData
Summary:
Every application in the ELP has an include search path, which is made up of the include directories in the transitive closure of all the dependencies of that app.

When doing buck `/include/` path mangling, we need to ensure we use the application that the specific include directory originates from.

To do this, we now store the local `include_dirs` in the `AppData` and use it for the name mangling.

This will also be used for generating an appropriate `include_lib` attribute for an arbitrary `FileId`, used to bring macro definitions into scope when completing them.

Reviewed By: TheGeorge

Differential Revision: D75454222

fbshipit-source-id: 15504bff25016e52223f344ac816266d1cdbfcbd