Add no_catch linter
Summary:
# Context
`try ... catch ... end` in Erlang has been available for a very long time, and
the old, simplistic `catch Expr` will be dropped from the language. In most
places where it is used, it carries with it some unwanted (and often unknown)
corner case behaviour, such as discarding the original error location,
conflating errors, exits, and throws, and mixing up thrown values with normally
returned values in a way that makes it very hard to see what the author
intended, and whether the implementation really does what they hoped.
Maintainance of such code is very hard.

Starting with Erlang OTP 28, it will be possible to enable warnings for use of
old-style `catch` expressions via the `warn_deprecated_catch` flag. The
old-style catch will be deprecated and removed in a future version of OTP (most
likely OTP 29 or OTP 30). Therefore, it is recommended to avoid switching to
`try ... catch ... end`.

The linter could later on be accompanied by a quick-fix.

Reviewed By: michalmuskala, alanz

Differential Revision: D74881343

fbshipit-source-id: 609d809dcb4da62b649f82344878d9816a386b71