2/n: Add completion for some well-known macros
Summary:
The first step is to add a table of macros to complete, and to provide completion for them.

It does not yet add the missing include attribute, nor does it provide arity/argument completion.

Later in the stack we will generate the table from the header files in the project, this allows us to get the machinery working properly first.

Reviewed By: robertoaloi

Differential Revision: D66368091

fbshipit-source-id: bb58b946deaab5a24fbfb39dd053a3c517c5ffc9