1/n: Add additional_edit field to Completion type
Summary:
This diff stack adds the ability to add an include(_lib) edit when completing a macro name known in the context.

This diff sets the stage, by adding a new field to the `Completion` type to capture the text of the required additional import if needed.  It does nothing else, the field is set to `None` everywhere. And provisionally has type `String`, this will change later in the stack.

Reviewed By: robertoaloi

Differential Revision: D66367892

fbshipit-source-id: 345647bfa6003ac0e6a1c72f93dc283558084df0