4/n: Improve auto-include attribute insertion
Summary:
When a well-known macro needs a specific include attribute to work, do not add it to the completion item if it is already present.

Also, add the include attribute after the module header.

The algorithm for choosing the location of the include attribute could probably do with some love, but that can be a follow up activity.

Reviewed By: robertoaloi

Differential Revision: D66370028

fbshipit-source-id: 53b005901611a11fddb631e3b7302f1919f73e5a