3/n proof of concept addtional edit to add include in completion
Summary:
The next step in this stack is to confirm that we can send the "add import" edit to the LSP client in the `additional_text_edits` edits field for the completion result, and have it do the necessary.

This is a bare bones implementation, it hard codes the insert location to the very top of the file, and does not check if the include attribute already exists.

These are dealt with in the next diff in the stack

Reviewed By: robertoaloi

Differential Revision: D66369471

fbshipit-source-id: a42b60289aebe40f14d31f28ae1c44989776c059