6/n Index of macro defines
Summary:
Create an index of all macros defined in `.hrl` files.
These are stored in a trie structure, based on the name/arity of the macro, so we can easily query it in a completion scenario. It stores the original macro definition and file, per macro name.

This diff just builds the index, it will be used in subsequent ones.

Reviewed By: TD5

Differential Revision: D66597868

fbshipit-source-id: 9ef88f94ce2bce5406c8ba0d540deb5c84a32da0