Performance improvements for project model
Summary:
The process of building a buck project model is slow because it explicitly walks the tree of buck2 target dependencies to construct include paths for each.

This diff uses a different approach, where it keeps a per-project mapping from the text of the `-include` or `-include_lib` argument to the file it resolves to.

This is a much shorter process, and drops the initial startup time to what it was before the detailed buck project model.

Note: it currently pre-computes the include map on initial load, and does not update it if new include files are added.  Because of the way the map is stores, this should be straightforward to implement.

Reviewed By: TheGeorge

Differential Revision: D76727805

fbshipit-source-id: 2b2f141a215abbddb9ba1b6ad403d7dff316ad5f