Avoid file system operations during buck target loading
Summary:
`.is_file` checks the file on disk - we don't need to do that. Since we're dealing with headers they'll be either and `.hrl` file or a directory - it's enough to check for the extension.

This still retains the file system access in `update_mapping_from_path` but reduces some of the utf8 shuffling done in there, but keeping everything in the realm of utf8 paths/

Reviewed By: alanz

Differential Revision: D75881672

fbshipit-source-id: 61f7c410d7ee058d1b3a63b570c46da13c7531e9