Add failing test for call hierarchy
Summary:
Demonstrate that when we request incoming calls for a function that is also used in a macro definition, we get no call hierarchy at all.

This is because the loop processing the references to the function uses `?`, so if any one reference process fails, no results are returned at all.

This is fixed in the next diff

Reviewed By: TD5

Differential Revision: D77230436

fbshipit-source-id: fd6224e29f448f2cf80dedb1d080d6bb47cc6a64