BinaryStringToSigil: Ignore multi-line binary strings (#104)
Summary:
In the current implementation, the "binary string to sigil" fix will create syntactically invalid code when applied to multi-line binary strings. This PR makes it so that multi-line binary strings are ignored for this diagnostic.

Currently, the LSP will suggest replacing this:

```erl
<<"Hello everyone! "
  "Goodbye!">>
```

With this:

```erl
~"Hello everyone! "
 "Goodbye!"
```

This is not valid Erlang syntax as far as I've been able to tell, and I don't think there is any good equivalent "sigil syntax" variant for multi-line strings.

Let me know what you think! Thanks

Pull Request resolved: https://github.com/WhatsApp/erlang-language-platform/pull/104

Reviewed By: ilya-klyuchnikov, TD5

Differential Revision: D76633464

Pulled By: alanz

fbshipit-source-id: da395d48396abbc640413bfb56f55e2fee2c44bf