Add an elp_development LLM guide
Summary:
Experiment, add a file describing the conventions used in the ELP codebase.

Note that the contents were initially generated by adding specific files to the context and asking the AI to update it based on that.

Then beaten in to shape to prevent the new diagnostic code being added in the wrong place, or replacing an existing one.

Reviewed By: robertoaloi, TD5

Differential Revision: D77308986

fbshipit-source-id: 0ae272faac297b128e1c2f01c35d04049238b16e